refactor(api): reorganize result classes into service.result package
