+++ api/src/main/java/net/onelitefeather/pandorascluster/api/service/result/CreateLandResult.java
@@ -1,6 +1,7 @@
-package net.onelitefeather.pandorascluster.api.service;
+package net.onelitefeather.pandorascluster.api.service.result;
 
 import net.onelitefeather.pandorascluster.api.land.Land;
+import net.onelitefeather.pandorascluster.api.service.LandService;
 
 /**
  * Outcome of {@link LandService#createLand}. Leaves room for future
